export to csv as function
